Devlop

+++ cubemall-product/src/main/java/com/kkb/cubemall/product/vo/Attr.java
@@ -1,6 +1,3 @@
-/**
-  * Copyright 2021 json.cn 
-  */
 package com.kkb.cubemall.product.vo;
 
 import lombok.Data;
@@ -8,7 +5,7 @@ import lombok.Data;
 @Data
 public class Attr {
 
-    private int attrId;
+    private Long attrId;
     private String attrName;
     private String attrValue;
