+++ cubemall-product/src/main/java/com/kkb/cubemall/product/vo/BaseAttrs.java
@@ -1,15 +1,18 @@
 /**
-  * Copyright 2021 json.cn 
-  */
+ * Copyright 2019 bejson.com
+ */
 package com.kkb.cubemall.product.vo;
 
 import lombok.Data;
 
+
 @Data
 public class BaseAttrs {
 
-    private int attrId;
+    private Long attrId;
     private String attrValues;
     private int showDesc;
+
+
 
 }
